turtle flip coin exercise
The turtle begins in the center of the screen.
Flip a coin. If its heads then turn to the left 90 degrees. If its tails then turn to the right 90 degrees.
Take 50 steps forward.
If the turtle has moved outside the screen then stop, otherwise go back to step 2 and repeat