creating sequence 3n_+1_
The rule for creating the sequence is to start from some given number, call it n, and to generate the next term of the sequence from n, either by halving n, whenever n is even, or else by multiplying it by three and adding 1 when it is odd. The sequence terminates when n reaches 1.